Add files via upload
This is a basic Springboot project that I have created. This is not connected to any database. To test this project I have used Postman(It has the ability to make various types of HTTP requests like GET, POST, PUT, PATCH). I havent connected it to database yet.
It stores the data as an array of list.